+++ services/settlement-service/src/main/java/com/northstar/settlement/domain/SettlementResult.java
@@ -1,8 +1,6 @@
 package com.northstar.settlement.domain;
 
-import java.math.BigDecimal;
-
-public record SettlementResult(BigDecimal coveredAmount,
-        BigDecimal deductibleApplied, BigDecimal depreciation,
-        boolean cappedAtLimit, BigDecimal settlementAmount) {
+/** Legacy {@code Settlement} bean values, kept as {@code double} like the source. */
+public record SettlementResult(double coveredAmount, double deductibleApplied,
+        double depreciation, boolean cappedAtLimit, double settlementAmount) {
 }
